Introduce use of symfony 5.x, flex, docker-compose (#174)
Add sensitive default for project dir
Install blackfire agent
Remove public bundle
Ensure blackfire configuration is present
Revise default variable assignments
Revise Symfony requirement default mode
Start migration of application from Symfony 3.x to Symfony 4.x
Add documentation paragraph about schema migration
Fix typo in information message related to database migration
Add shortcut function to run composer commands
Fix documentation of function called to migrate schema
Add help message to run PHP script
Upgrade Dockerfile to build PHP 7.4.x image
Delete dead code
Migrate application to Symfony 5.0
Fix class imports
Fix class imports
Ensure a single route is available
Implement missing method in authenticator
Fix compatibility of console events subscriber for Console component API
Remove directory containing configuration files for previous version of Symfony
Remove duplicated routes
Rename user to member
Replace deprecated class in kernel
Import configuration files for installed packages
Invert membership and member entity configuration
Declare service definition of member repository
Replace deprecated class usage in kernel (routing)
Fix class import in token authenticator
Remove redundant constructor from member repository
Remove deprecated entries from VCS ignore list
Add class ensuring consistency of statuses (for full text and status id)
Install PHPUnit and Safe libraries
Remove tests failing because of framework changes
Import configuration of xdebug for PHP cli
Let the development web server add CORS headers
Replace codeship parameters
Replace travis parameters
Do not build bootstrap file anymore
Downgrade PHP version of travis builds
Fix travis source test parameters file
Remove wrong configuration from travis and codeship files
Fix password used by codeship to run tests
Revise shortcut function executed to run tests
Remove unnecessary import
Add documentation note about where MySQL database password comes from in continuous integration
Fix retrieval of rabbitmq virtual host from local configuration file
Add DQL extensions to access native functions
Ensure there is at least an access token in the database
Install amqp PHP extension
Ensure messages to fetch member statuses can be dispatched
Rename dispatcher of messages to fetch member statuses
Fix retrieval of messenger virtual host
Introduce handler of message consumed to fetch member statuses
Add credentials to Fetch member statuses message
Fix class import
Limit range of ip addresses of press-review-network
Fix display of password passed to rabbitmq
Upgrade PHP version
Declare message handler instead of message consumer
Revise remote debug host to comply with docker network of php container
Replace message consumer with message handler depending on messenger component
Install intl extension for PHP depending on libicu-dev package and amqp extension for PHP
Reformat code
Revise type hinting for some methods
Fix definition of API accessor service
Revise configuration of script to initialize apache virtual host
Typehint properties of list controller
Revise entrypoint of apache container
Introduce project configuration file for docker-compose
Ensure entryoint of worker container is blocking
Ensure worker belongs to press review network
Ensure cache server keeps appending values
Ensure entrypoint of worker container is blocking
Add migration file to remove group and role tables
Revise commands to migrate schema
Add distribution configuration file for the application
Remove hash from default extensions file
Ensure building the worker does not fail because of the extensions configuration file
Remove extra step consisting in reinstalling sockets extension for PHP
Let web container depends on worker container
Prevent unique token from failing application
Fix call to substring
Add missing docblock
Revise press review subnet
Set application environment variable to prod
Ensure consumer key and secret are nullable
Revise error message logged when authentication data are incorrect
Remove application environment variable set to prod
Add amp extension
Remove addition of configuration file by docker-compose
Rename shortcut functions executed to handle messages
Remove function executed to list PHP extensions
Remove command to set up amqp exchanges
Rename message producer to message dispatcher
Refactor commands depending on app/console so that they depend on bin/console
Revise command to list queues
Add shortcut command to run worker
Revise shortcut command to set permissions of cache and log directories
Replace app/logs with var/logs
Refactor commands depending on app/console so that they depend on bin/console
Rely on docker-compose to execute commands
Migrate environmment variable declared set application environment
Fix shortcut functions executed to dispatch messages
Set default memory limit to 64M when handling messages
Remove redundant command
Add shortcut command to run the whole stack
Revise supervisor configuration
Revise function setting permissions
Revise function setting permissions to clear cache for development environment
Execute message consumption in detached mode with a time limit
Dispatch messages in detached mode

+++ provisioning/containers/apache/tasks/setup-virtual-host.sh
@@ -1,27 +1,33 @@
 #!/bin/bash
 
-cd /etc/apache2
+function init_virtual_host() {
+    cd /etc/apache2
 
-ln -s /templates/sites-enabled/press-review.conf /etc/apache2/sites-available
+    if [ ! -e /etc/apache2/sites-available ];
+    then
+        ln -s /templates/sites-enabled/press-review.conf /etc/apache2/sites-available
+    fi
 
-# Disable default virtual host
-a2dissite 000-default.conf
+    # Disable default virtual host
+    a2dissite 000-default.conf
 
-# Enable rewrite module
-a2enmod rewrite
+    # Enable rewrite module
+    a2enmod rewrite
 
-# Enable headers module for CORS
-a2enmod headers
+    # Enable headers module for CORS
+    a2enmod headers
 
-# Enable HTTP2
-a2enmod http2
+    # Enable HTTP2
+    a2enmod http2
 
-cd /etc/apache2
-ln -s `pwd`/sites-available/press-review.conf `pwd`/sites-enabled
+    cd /etc/apache2
+    ln -s `pwd`/sites-available/press-review.conf `pwd`/sites-enabled
 
-rm `pwd`/mods-enabled/deflate.conf
-ln -s /templates/mods-enabled/deflate.conf `pwd`/mods-enabled
+    rm `pwd`/mods-enabled/deflate.conf
+    ln -s /templates/mods-enabled/deflate.conf `pwd`/mods-enabled
 
-/etc/init.d/blackfire-agent restart
+    /etc/init.d/blackfire-agent restart
 
-apache2-foreground &
+    apache2-foreground &
+}
+init_virtual_host
